18th commit
数据标准化，学生能力指标映射

+++ src/main/java/com/klaus/workserviceimpl/ScoreWRWorkerServiceImpl.java
@@ -1,7 +1,10 @@
 package com.klaus.workserviceimpl;
 
 import java.io.File;
+import java.util.List;
 
+import com.klaus.dao.FileUploadStatementDAO;
+import com.klaus.dao.TempScoreFileDAO;
 import com.klaus.dao.WorkerDAO;
 import com.klaus.factory.MyBeansFactory;
 import com.klaus.utils.StaticData;
@@ -26,7 +29,23 @@ public class ScoreWRWorkerServiceImpl implements WorkerService {
 			return;
 
 		}
+		
+		System.out.println("1111111111111111");
+		
+		
+		FileUploadStatementDAO fileUploadStatementDao=(FileUploadStatementDAO)MyBeansFactory.getBeans("fileuploadstatementdao");
+		
+		int count=fileUploadStatementDao.getFilesCount(1);
+		
+		if(count!=0){
+			
+			return ;
+			
+		}
 
+		
+		System.out.println("222222222222222222");
+		
 		WorkerDAO service = (WorkerDAO) MyBeansFactory.getBeans("workerdao");
 
 		int flag = service.working(StaticData.StringData.ScoreWorker);
@@ -37,20 +56,45 @@ public class ScoreWRWorkerServiceImpl implements WorkerService {
 
 		}
 
+		
+		System.out.println("3333333333333333");
+		
 		service.startWork(StaticData.StringData.ScoreWorker);
 
+		
+		TempScoreFileDAO tempScoreFileDa=(TempScoreFileDAO)MyBeansFactory.getBeans("tempscorefiledao");
+		List<String> files=tempScoreFileDa.getTempFiles();
+		
+		
+		System.out.println("4444444444444444444");
+		
+		ExcelDBService scoreExcel = (ExcelDBService) MyBeansFactory.getBeans("scoreexcelDBServiceimpl");
+		
 		for (int i = 0; i < length; i++) {
 
+			
+			System.out.println("55555555555555555555555");
+			
 			String filePath = StaticData.StringData.FileScorePath + fileList[i];
 
-			File file = new File(filePath);
+			if(files==null||files.size()==0||files.indexOf(fileList[i])==-1){
+				
+				
+				File file = new File(filePath);
+				
+				System.out.println("66666666666666666666");
+				
+				scoreExcel.saveData(filePath);
 
-			ExcelDBService scoreExcel = (ExcelDBService) MyBeansFactory.getBeans("scoreexcelDBServiceimpl");
-			scoreExcel.saveData(filePath);
-
-			file.delete();
-
+				
+				file.delete();
+				
+			}
+			
 		}
+		
+		
+		System.out.println("777777777777777777");
 
 		service.finishWork(StaticData.StringData.ScoreWorker);
